Added initial values example

+++ pybma/simulation.py
@@ -103,7 +103,7 @@ def simulate(qn,steps=10,initial_values=None):
             t0[var] = System.Int32(int(0))
     else:
         for var in initial_values.keys():
-            t0[var] = System.Int32(initial_values[var])
+            t0[System.Int32(var)] = System.Int32(initial_values[var])
     
     t0 = python_dict_to_fsharp_map(t0)
